CHROMIUM: sysrq: Added Chrome OS specific 'x' key
On Chrome OS systems the magic sysrq x key will cause the following to
happen:
- show processes in blocked state
- emergency file sync
- crash the system

This key combination can be used in the field by users to cause the
system to crash so that crash dumps can be collected to help debug UI
freeze issues.

BUG=chromium-os:33249
TEST=First "echo 0x1000 > /proc/sys/kernel/sysrq", then do the two tests
below:
1. Alt-F10-x should cause the system to crash. Looks for kcrash logs
under /var/spool/crash when the system reboots.
2. All other magic sysrq operation aside from alt-F10-x should be
ineffective.

This is a squash of the following commits:
  CHROMIUM: sysrq: Added Chrome OS specific 'x' key
  CHROMIUM: sysrq: fix Chrome OS specific 'x' key handler to work on ARM

Change-Id: I62d0a7f0028c072928be0125da4e069bafa5d88c
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29316
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>

[rebase44(filbranden): Squashed and reordered commits.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>