CHROMIUM: go2001: Don't immediately wait for replies to (most) messages.
To allow for more parallelism, don't wait for messages to be replied to
from the point of where they are queued. Instead, queue them and return
immediately.

To be able to do this, we need an ability to queue messages and not wait
for them to be sent to hardware (while the ring buffer is full), which
requires pending message memory to be managed separately. Switch to
allocating their buffers dynamically. To make this more optimal, use
kmem_cache.

Messages are allocated when queued to be send to hardware, and freed after
copied to the ring buffer, or on error, if still pending at that point.

Given that VM_EMPTY_BUFFER messages are scheduled in the interrupt context,
their memory has to be allocated beforehand. Do this by allocating in
buf_prepare callback and storing the memory on source go2001_buffer.

Not waiting for VM_EMPTY_BUFFER messages introduces an additional
complication on resolution change. When receiving a resolution change
notification in reply from hardware, we may have additional messages
pending already, which will be processed and returned to us from hardware
with a request to reprocess them after handling the resolution change.
To handle this, move all buffers that hardware requested to be reprocessed
to a new resume queue. After the resolution change sequence is finished,
move them back in front of the source queue for reprocessing. This also
requires us to allocate another message buffer for reprocessing at that
time.

Finally, waiting for context to be finished now involves waiting for the
pending message queue to become empty (i.e. messages are moved to the
outgoing message ring buffer), followed by waiting for replies to all
these messages to arrive from hardware.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chromium:473434
TEST=vda/veatest,Hangouts
Reviewed-on: https://chromium-review.googlesource.com/289983
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Trybot-Ready: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>

(cherry picked from commit bb6a04c337135df7fb0396a6c33efd939429e8f1)
Signed-off-by: Fletcher Woodruff <fletcherw@chromium.org>

Change-Id: I5da39070fd7ee1b80eec39ca46387b96cbfc1e97
Reviewed-on: https://chromium-review.googlesource.com/1650181
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>