CHROMIUM: iwl7000: mac80211: introduce capability flags for VHT EXT NSS support
Depending on whether or not rate control supports selecting
rates depending on the bandwidth, we can use VHT extended
NSS support. In essence, this is dot11VHTExtendedNSSBWCapable
from the spec, since depending on that we'll need to parse
the bandwidth.

If needed, also set/clear the VHT Capability Element bit for
this capability so that we don't advertise it erroneously or
don't advertise it when we actually use it.

Change-Id: Ic97b29e37fe5d2afef88b767d7dd28c56a50d2bf
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 99907e7a866f7c93b0400224bf5f141d1c1d01f3