UPSTREAM: PM / Domain: Add struct device to genpd
The power-domain core would be using the OPP core going forward and the
OPP core has the basic requirement of a device structure for its working.

Add a struct device to the genpd structure. This doesn't register the
device with device core as the "dev" pointer is mostly used by the OPP
core as a cookie for now and registering the device is not mandatory.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Acked-by: Ulf Hansson <ulf.hansson@linaro.org>

BUG=b:80592702
TEST=Build and boot

Change-Id: I4578538a9c7cd90a75e3f8ca1d2fff611986b408
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
(cherry picked from commit 401ea1572de944df548a13eded82339491a739ff)
Reviewed-on: https://chromium-review.googlesource.com/1094832
Commit-Ready: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>