CHROMIUM: mfd: cros_ec_pd_update: Add driver for updating of PD device firmware
Add a driver for auto-updating of PD devices. The driver currently performs
the following task: Upon receiving a host event, probe every PD port on the
EC, and send commands to update the PD device firmware if mismatched
firmware is found.

BUG=chromium:785506
TEST=Using the whole suite, check on eve,cyan,glimmer AIDA64 works.
Check on pixel 2 usb-pd-charger exits.

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Squash and syntax fixes of
- d1528e626e05 ("CHROMIUM: mfd: cros_ec_pd_update: refresh the power supply state")
- b9a7e5b751a9 ("CHROMIUM: mfd: cros_ec_pd_update: Add driver for updating of PD device firmware")

Change-Id: Ia4905f3586237b2379520860566884820b56daec
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>