CHROMIUM: iwl7000: mvm: support HE TXing with MCS10 and MCS11
Allow setting an HE rate of MCS10 and MCS11 for TXing.

Note that this patch does not affect the rate scaling
whatsoever, and the driver will still not reach HE
rates without setting it as a fixed rate. However, as
this is expected to be done by the FW, it is good
enough for now.

Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
iwl7000-tree: bbb17628b2e86dae6d788c1620f6cc977a616bbc