UPSTREAM: drm/amd/display: properly turn off unused mpc before front end programming
MPCC_OPP_ID must be programmed to 0xf to properly turn off the mpcc.
However the software state of the mpcc must keep track of the opp that
the mpcc is attached to for reset to properly happen. This is kinda
hacky right now, but a good solution may involve a lot of work.

BUG=b:65489307
TEST=boot Kahlee, run graphics_SanAngeles

Signed-off-by: Eric Yang <Eric.Yang2@amd.com>
Reviewed-by: Tony Cheng <Tony.Cheng@amd.com>
Acked-by: Harry Wentland <Harry.Wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 1674d35bf57b0546577b87af266e45de3ccf45c0)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: Id2a09cef9d32857615a3617c8ac9d40ab977a52e