CHROMIUM: iwl7000: mvm: temp disable bar frames 11ax client test mode
The FW code for 11ax client assumes no BAR frames or
control frames in general are sent from the driver.
This should be true after the agg manager offload task
is complete. This is a w/a so for now so we can test 11ax trigger
flows until the real solution is done.

This patch disables BAR frame sending from the driver
when the 11ax softap client test mode is active.

Signed-off-by: Jonathan Afek <jonathanx.afek@intel.com>
iwl7000-tree: 9d066a8d6187441a8a0feb7435982832fe15414d