UPSTREAM: tpm_tis_spi: Use DMA-safe memory for SPI transfers
The buffers used as tx_buf/rx_buf in a SPI transfer need to be DMA-safe.
This cannot be guaranteed for the buffers passed to tpm_tis_spi_read_bytes
and tpm_tis_spi_write_bytes. Therefore, we need to use our own DMA-safe
buffer and copy the data to/from it.

The buffer needs to be allocated separately, to ensure that it is
cacheline-aligned and not shared with other data, so that DMA can work
correctly.

Fixes: 0edbfea537d1 ("tpm/tpm_tis_spi: Add support for spi phy")
Cc: stable@vger.kernel.org
Reviewed-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Signed-off-by: Alexander Steffen <Alexander.Steffen@infineon.com>
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>

BUG=chromium:957203
TEST=Build and boot

Change-Id: Ia2d68ddbcbd69f59bf8235c044e857e399c0d75f
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 6b3a13173f23e798e1ba213dd4a2c065a3b8d751)
Reviewed-on: https://chromium-review.googlesource.com/957439
Reviewed-by: Andrey Pronin <apronin@chromium.org>