UPSTREAM: drm/msm: add special _get_vaddr_active() for cmdstream dumps
Prep work for adding a debugfs file that dumps just submits which
trigger hangs/faults.  In this case the bo may already be in the
MADV_DONTNEED state, but will be still on the active list (since
the submit hasn't completed yet).  So the normal check that the
bo is in the WILLNEED state does not apply.  (But of course the bo
should definitely not be in the PURGED state!)

Signed-off-by: Rob Clark <robdclark@gmail.com>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I711a9d081471726cd660134919d47f2a46de175c
Signed-off-by: Archit Taneja <architt@codeaurora.org>
(cherry picked from commit fad33f4b1073a423a9ecd2a335de05e7a3ec1f37)
Reviewed-on: https://chromium-review.googlesource.com/859220
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>