UPSTREAM: drm/msm: preserve IOVAs in submit's bo table
We need this if we want to dump the submit after cleanup (ie. from hang
or fault).  But in the backoff/unpin case we want to clear them.  So add
a flag so we can skip clearing the IOVAs in at cleanup.

Signed-off-by: Rob Clark <robdclark@gmail.com>

BUG=chromium:795946
TEST=Build and boot

Change-Id: Ibdacd180c0d5a48f313c6e966c51451c53e05f61
Signed-off-by: Archit Taneja <architt@codeaurora.org>
(cherry picked from commit 544a32dbf05888ec012868db2e813233078bbdfb)
Reviewed-on: https://chromium-review.googlesource.com/859223
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>