UPSTREAM: drm/i915/execlists: Split insert_request()
In the next patch we will want to reinsert a request not at the end of
the priority queue, but at the front. Here we split insert_request()
into two, the first function retrieves the priority list (for reuse for
unsubmit later) and a wrapper function to insert at the end of that list
and to schedule the tasklet if we were first.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Michał Winiarski <michal.winiarski@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20170916204414.32762-3-chris@chris-wilson.co.uk
(cherry picked from commit  27606fd87895c3b45072a796caead8055a2f4b94  
git://anongit.freedesktop.org/drm-intel drm-intel-next-queued)
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>

BUG=chromium:783394
TEST=CNL boots to graphics

Change-Id: I20deae368472f2e631ee4123a80260d0bafb2e6a