CHROMIUM: iwl7000: mvm: add CPU DMA latency requirement
In high througput scenarios, the CPU should only be allowed to enter
C-states whose exit latency are not too high, because this can impact
DMA latency.  To achieve this, use the PM QoS framework to require low
DMA latency when we are in high-load situations, as measured by TCM.

The sweet spot in a specific system has been found to be 60 usecs
maximum latency, which achieves the best possible power-saving while
still not hurting throughput.

Since we don't have a formal way of calculating the theoretical
maximum latency requirement yet, use 60 usecs, which was found to be
the sweet spot in a specific system that achieves the best possible
power-saving while still not hurting throughput.

Signed-off-by: Rajneesh Bhardwaj <rajneesh.bhardwaj@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 4bd6ea948fb592d5f6a46f5aaa454924620ba8c8