UPSTREAM: Bluetooth: avoid recursive locking in hci_send_to_channel()
Mart reported a deadlock in -RT in the call path:
  hci_send_monitor_ctrl_event() -> hci_send_to_channel()

because both functions acquire the same read lock hci_sk_list.lock. This
is also a mainline issue because the qrwlock implementation is writer
fair (the traditional rwlock implementation is reader biased).

To avoid the deadlock there is now __hci_send_to_channel() which expects
the readlock to be held.

Fixes: 38ceaa00d02d ("Bluetooth: Add support for sending MGMT commands and events to monitor")
BUG=chromium:786148
TEST=B oot multiple times and check if deadlock occurs

Change-Id: Ib7b611e86a0776058ccb533eca5097ac4b86ed8e
Reported-by: Mart van de Wege <mvdwege@gmail.com>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit a9ee77af751f)
Reviewed-on: https://chromium-review.googlesource.com/777642
Reviewed-by: Dylan Reid <dgreid@chromium.org>