UPSTREAM: iommu/iova: Optimise rbtree searching
Checking the IOVA bounds separately before deciding which direction to
continue the search (if necessary) results in redundantly comparing both
pfns twice each. GCC can already determine that the final comparison op
is redundant and optimise it down to 3 in total, but we can go one
further with a little tweak of the ordering (which makes the intent of
the code that much cleaner as a bonus).

Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Tested-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Zhen Lei <thunder.leizhen@huawei.com>
Tested-by: Nate Watterson <nwatters@codeaurora.org>
[rm: rewrote commit message to clarify]
Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 2070f940a6d5148cf2df0d0087ff0a64d9f15237)

BUG=b:110260554
TEST=Build kernel for cheza

Change-Id: Iacb1f122f79e35bdce13be78fb5070f375b7c480
Signed-off-by: Vivek Gautam <vivek.gautam@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/1125859
Reviewed-by: Tomasz Figa <tfiga@chromium.org>