bluetooth: "Fix" sync of commands for adv & scan
This is a very ugly hack, presented here in a very ugly way. The ugly
way is basically how we deliver this only for eve and not affect other
devices. The ugly hack is basically how we work aroung bluez's complete
lack of synchronization. Let me explain. No, there is not time. Let me
summarize: bluez has a boolean value for "are we scanning" which it
updates inside the command-complete event handler. It also checks this
in many places as a way to tell if the chip is or is not scanning. The
issue is that some chips (like BCM's) let you disable something already
disabled and do not raise a fuss. Other chips (like Intel's 7265) do
raise a fuss if you try to disable scanning if it is already off. So
why the issue? If the chip raises an error, bluez loses its marbles and
gets out of sync with the chip. Basically it thinks the chip *IS*
scanning while chip is not. There is no way out of this since any
attempt to tell bluez to start scanning causes it to just reply with
"already am" and any attempt to stop scanning sends the "scan disable"
command to the chip and produces the same kind of error that got us
into this mess. So how do we fix this? There isn't really a simple or
elegant way. But the basic issue is that while a command to enable or
disable scanning is in flight, any code in the stack will think
scanning is still in the old state, while it may already be in the new
state. This might cause more commands to be enqueued based on this
errorneous information. To solve this we'd like to always have correct
information. Sadly this is not easy. There is also a command queue,
which at any point in time may have many command enqueued that affect
scanning state. There are multiple ways to insert commands into it,
and we do not want to instrument them all. So what do we do? First we
add a second boolean bit for "scan state change command in flight".
We set this when we *SEND TO THE CHIP* a command that affects scan
state and clear it when a reply arrives. We also change all places
that check the "are we scanning" state to properly consider this new
variable as well. This will not solve the issue of many commands
queued already. So it is still possible that while a "disable" was
queued, another "disable" may be queued. To work around this we also
track the current state in the lowest layer that actually sends
commands to the chip and simply drop ones that would transition us to
the same state that we're already in. The situation is actually
weirder because there is a "restart" func in bluez all of whose
callers seem to assume that it atomically toggles scanning off and back
on if it is already on. No such atomicity exists in reality. Our hack
covers up for this in most cases. Oh, and this same exact logic applies
to "are we advertising" boolean and we treat it the same way.

This really solves nothing, but it works around the issue well enough
for our purposes. A proper fix would be a separate queue for state
changes for "adv" and "scan" with proper suport for atomic "restart"
requests. This queue would then submit actual commands to the chip's
queue. since nobody else would do that, it would have no races. We have
no time to produce such fixes, so for now you get to enjoy this dirty
hack.

This hack is based on initial work by josephsih@ who first figured out
the proper place to filter all the low level commands as they are sent
to the chip and the proper way to drop them.

BUG=b/66502139,b:128319712
TEST=it did not blow up my chromebook and long-endirance scan test ran
14+ hours (bt scan in bluetoothcli with no stop)

Change-Id: I64a4d9be52ff4472acb8f81cffef69b1ef33ac80
Reviewed-on: https://chromium-review.googlesource.com/677863
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Commit-Queue: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1617802
Commit-Ready: Yoni Shavit <yshavit@chromium.org>
Tested-by: Yoni Shavit <yshavit@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>