UPSTREAM: drm/amdgpu: Reserve shared memory on VRAM for SR-IOV
SR-IOV need to reserve a piece of shared VRAM at the exact place
to exchange data betweem PF and VF. The start address and size of
the shared mem are passed to guest through VBIOS structure
VRAM_UsageByFirmware.

VRAM_UsageByFirmware is a general feature in VBIOS, it indicates
that VBIOS need to reserve a piece of memory on the VRAM.

Because the mem address is specified. Reserve it early in
amdgpu_ttm_init to make sure that it can monoplize the space.

BUG=b:65489307
TEST=boot Kahlee, run graphics_SanAngeles

Signed-off-by: Horace Chen <horace.chen@amd.com>
Reviewed-by: Christian Knig <christian.koenig@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit a05502e5cfa9abe17a16592be82c2f5692c91f35)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I1b42f1033237c16b0643e20f9b2224eb39836c01
Reviewed-on: https://chromium-review.googlesource.com/799457
Commit-Ready: Dominik Behr <dbehr@google.com>
Tested-by: Dominik Behr <dbehr@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>