CHROMIUM: Bluetooth: mgmt: add additional HCI cmd to add_advertising
When a process registers and unregisters advertisements consecutively,
it is likely that an advertisement with the same adv data and adv
response data is unregistered and then registered again immediately.
In the case of a short period with a single advertisement, it is
possible that the adv related data is still in the controller and
there is no need to update the controller with the new data. Hence,
there are no HCI commands built in the request. However, it was
originally treated as an error.

A simple work around again is to insert a whatever HCI command in the
request command queue to prevent the queue from being empty.

There is supposed to be a better way to handle this situation.

BUG=chromium:760792
TEST=None

Change-Id: I17196f3b0ca0684e556307905a651ba48da11c31
Signed-off-by: Joseph Hwang <josephsih@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/694949
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1648872
Commit-Ready: Yoni Shavit <yshavit@chromium.org>
Tested-by: Yoni Shavit <yshavit@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>