FROMLIST: drm/i915/cnl: WaSendPushConstantsFromMMIO / WaDisableGatherAtSetShaderCommonSlice
WaSendPushConstantsFromMMIO: "If not using RS, we must send two
MMIO registers at context create to trigger push constants at
3D primitive"

There is almost no documentation on Spec or wa_database about
this WaSendPushConstantsFromMMIO. So we also found out in another
place that WaSendPushConstantsFromMMIO was only adding
COMMON_SLICE_CHICKEN2 in a white list.

So we looked to the programming notes of COMMON_SLICE_CHICKEN2
and we notice that this bit 12 is marked in association with
2 other MMIO registers for SKL+. Apparently for SKL+ we should
check few MMIOs to decide for set or reset of this bit 12.

Also "gather" is related to gather and packing of constants elements
into "push constants".

Mesa has no plans of using RS so there is no need to whitelist
and we only need to initialize the context image with that bit clear.

v2: Move to cnl_init_workarounds as Chris suggested.
    Add both Wa names and improve commit message as Rafael suggested.

Cc: Rafael Antognolli <rafael.antognolli@intel.com>
Cc: Mika Kuoppala <mika.kuoppala@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Tested-by: Rafael Antognolli <rafael.antognolli@intel.com>
(git am from https://patchwork.kernel.org/patch/9988065/)
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>

BUG=chromium:783394
TEST=CNL boots to graphics

Change-Id: I1c6a6485c353c7101c7c1ac61edf7e595b2ce361