UPSTREAM: drm/i915: Pin fence for iomap
Acquire the fence register for the iomap in i915_vma_pin_iomap() on
behalf of the caller.

We probably want for the caller to specify whether the fence should be
pinned for their usage, but at the moment all callers do want the
associated fence, or none, so take it on their behalf.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20171009084401.29090-1-chris@chris-wilson.co.uk
(cherry picked from commit b4563f595ed44514467cfb3566b8b6519f328354)

BUG=chromium:783394
TEST=CNL boots to graphics

Change-Id: Id91ea4b1fd35cf4356439e06d1aae5245c319749
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/843430
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>