UPSTREAM: fscrypt: new helper function - fscrypt_require_key()
Add a helper function which checks if an inode is encrypted, and if so,
tries to set up its encryption key.  This is a pattern which is
duplicated in multiple places in each of ext4, f2fs, and ubifs --- for
example, when a regular file is asked to be opened or truncated.

Acked-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit d293c3e4e07334d761c88df9d68b3aa800a83dd9)

BUG=b:214630925
TEST=compile

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I1e6f1e1763d40889520ef6cedf09e92ee574fe44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3435672
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>