UPSTREAM: fscrypt: new helper function - fscrypt_file_open()
Add a helper function which prepares to open a regular file which may be
encrypted.  It handles setting up the file's encryption key, then
checking that the file's encryption policy matches that of its parent
directory (if the parent directory is encrypted).  It may be set as the
->open() method or it can be called from another ->open() method.

Acked-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit efcc7ae2c9172d9a7ae94afdaf066a7abf0b9a90)

BUG=b:214630925
TEST=compile

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I41df494d18971fa208105af633ca2ddb5b350ad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3435674
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>