UPSTREAM: fscrypt: new helper function - fscrypt_prepare_rename()
Introduce a helper function which prepares to rename a file into a
possibly encrypted directory.  It handles loading the encryption keys
for the source and target directories if needed, and it handles
enforcing that if the target directory (and the source directory for a
cross-rename) is encrypted, then the file being moved into the directory
has the same encryption policy as its containing directory.

Acked-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit 94b26f3672a0e41025104c7e46943917544e1c87)

BUG=b:214630925
TEST=compile

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I4d070db6f42618130b3cd17a99cbfef90e9770c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3435676
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>