UPSTREAM: of: reserved_mem: Accessor for acquiring reserved_mem
In some cases drivers referencing a reserved-memory region might want to
remap the entire region, but when defining the reserved-memory by "size"
the client driver has no means to know the associated base address of
the reserved memory region.

This patch adds an accessor for such drivers to acquire a handle to
their associated reserved-memory for this purpose.

A complicating factor for the implementation is that the reserved_mem
objects are created from the flattened DeviceTree, as such we can't
use the device_node address for comparison. Fortunately the name of the
node will be used as "name" of the reserved_mem and will be used when
building the full_name, so we can compare the "name" with the basename
of the full_name to find the match.

Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I8f8e2d921ef6e3ed92be66ef4aa16a8378510bf3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit eb297bc716ec2cb3147e6e99002e37058c65cba3)
Reviewed-on: https://chromium-review.googlesource.com/843566
Reviewed-by: Ben Chan <benchan@chromium.org>