UPSTREAM: dt-binding: soc: qcom: Add binding for rmtfs memory
This adds the binding for describing shared memory used to exchange file
system blocks between the RMTFS client and service. A client for this is
generally found in the modem firmware and is used for accessing
persistent storage for things such as radio calibration.

Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I9f441796453d222468908148f4e91f6f208bfb17
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 2501ec14048d24d98651d2204cdf9cb33a127199)
Reviewed-on: https://chromium-review.googlesource.com/843567
Reviewed-by: Ben Chan <benchan@chromium.org>