UPSTREAM: net: qrtr: Pass source and destination to enqueue functions
Defer writing the message header to the skb until its time to enqueue
the packet. As the receive path is reworked to decode the message header
as it's received from the transport and only pass around the payload in
the skb this change means that we do not have to fill out the full
message header just to decode it immediately in qrtr_local_enqueue().

In the future this change also makes it possible to prepend message
headers based on the version of each link.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I5ee9c210f43837af7ded4dfdd086164316d3f306
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit e7044482c8ac5081f4775063995647787d5082a4)
Reviewed-on: https://chromium-review.googlesource.com/843558
Reviewed-by: Ben Chan <benchan@chromium.org>