UPSTREAM: net: qrtr: Clean up control packet handling
As the message header generation is deferred the internal functions for
generating control packets can be simplified.

This patch modifies qrtr_alloc_ctrl_packet() to, in addition to the
sk_buff, return a reference to a struct qrtr_ctrl_pkt, which clarifies
and simplifies the helpers to the point that these functions can be
folded back into the callers.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I7edd236d90cefe6dd825fa3412e5dae53c1808b0
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 1a7959c76641674991ba3a49f25432f1e105391e)
Reviewed-on: https://chromium-review.googlesource.com/843559
Reviewed-by: Ben Chan <benchan@chromium.org>