UPSTREAM: net: qualcomm: rmnet: Implement bridge mode
Add support to bridge two devices which can send multiplexing and
aggregation (MAP) data. This is done only when the data itself is
not going to be consumed in the stack but is being passed on to a
different endpoint. This is mainly used for testing.

Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BUG=chromium:795946
TEST=Build and boot

Change-Id: Iebedf20bdb1d8a24d159ebb4b12189aaa60d5fbf
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 60d58f971c1077a0f2467b2d5bc38058df43a819)
Reviewed-on: https://chromium-review.googlesource.com/865549
Reviewed-by: Guenter Roeck <groeck@chromium.org>