UPSTREAM: drm/amdgpu: report more amdgpu_fence_info
Only for GFX ring. This can help checking MCBP feature.

The fence at the end of the frame will indicate the completion status.
If the frame completed normally, the fence is written to the address
given in the EVENT_WRITE_EOP packet. If preemption occurred in the
previous IB the address is adjusted by 2 DWs. If work submitted in the
frame was reset before completion, the fence address is adjusted by
four DWs. In the case that preemption occurred, and before preemption
completed a reset was initiated, the address will be adjusted with six
DWs

BUG=b:65489307
TEST=boot Kahlee, run graphics_SanAngeles

Signed-off-by: pding <Pixel.Ding@amd.com>
Reviewed-by: Monk Liu <monk.liu@amd.com>
Reviewed-by: Christian Knig <christian.koenig@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit e71de0766191d32648cf12dfb2f53f05e52b2dcc)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I925290714518f21a54f6f12a29bf203968808ebd
Reviewed-on: https://chromium-review.googlesource.com/806989
Commit-Ready: Dominik Behr <dbehr@google.com>
Tested-by: Dominik Behr <dbehr@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>