CHROMIUM: iwl7000: mvm: fix setting lq_color
Commit 6ba187969ede ("[BUGFIX] iwlwifi: mvm: add missing lq_color")
assigned the rate scaling directly to driver_data[0].
However, the conversion for the rate scaling is done later in
iwl_mvm_tx_reclaim.
This double conversion happened to work, since the macro isn't
masking the values and the already converted value was simply
or'ed with 0, so the second call actually didn't do anything.

The real issue is that the tid_data lq_color field was never
assigned. This happened since it was initialized when receiving
tx response for aggregation, which we no longer get for a000
devices.
Properly initialize it to avoid the double usage of the rate
scaling macro.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: f1b2ee8be99c1bf3ad554245bcd59015ce67e918