UPSTREAM: drm/atomic: Check for busy planes/connectors before setting the commit
We still want to fail with -EBUSY if a plane or connector is part of
a commit, even if it will be assigned to a new commit.

This closes a small hole left open where we should return -EBUSY.
It's not severe, because wait_for_dependencies and swap_state helpers
still block. But it should return -EBUSY and not stall.

BUG=b:65489307
TEST=boot Kahlee, run graphics_SanAngeles

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Fixes: 21a01abbe32a ("drm/atomic: Fix freeing connector/plane state too early by tracking commits, v3.")
Link: https://patchwork.freedesktop.org/patch/msgid/20171016132928.6498-2-maarten.lankhorst@linux.intel.com
Reviewed-by: Ville Syrjl <ville.syrjala@linux.intel.com>
(cherry picked from commit 080de2e5be2d4b396b6792cde4cb2048fcdb4ce9)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I79acf26720bedc0bc4f93ddafa602395bd84c9c7
Reviewed-on: https://chromium-review.googlesource.com/807042
Commit-Ready: Dominik Behr <dbehr@google.com>
Tested-by: Dominik Behr <dbehr@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>