UPSTREAM: dt-bindings: display: msm: update clk names
Now that drm/msm is converted over to use msm_get_clk() everywhere (that
matters), which handles falling back to looking for a clock with the
"_clk" suffix, we can remove "_clk" from the documentation so that new
dts files added do not include "_clk" in the name.

Previously we were doing this for the more recently upstreamed bindings
but not for (nearly) all.

Signed-off-by: Rob Clark <robdclark@gmail.com>
Acked-by: Rob Herring <robh@kernel.org>

BUG=chromium:795946
TEST=Build and boot

Change-Id: Ibbe6054fe2df807f6de802c05ec18be1c1fb7916
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 63d51e36b5add80bb2d94e67de3d643854faaf10)
Reviewed-on: https://chromium-review.googlesource.com/864783