UPSTREAM: irqchip/gic: Make quirks matching conditional on init return value
As it turns out, the IIDR is not sufficient to distinguish between GICv3
implementations when it comes to enabling quirks. So update the prototype
of the init() hook to return a bool, and interpret a 'false' return value
as no match, in which case the 'enabling workaround' log message should
not be printed.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>

BUG=chromium:795946
TEST=Build and boot

Change-Id: If9f3053135fc31b85160aff9bc971de7010c0b8e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 9d111d49106b61f5a652d5418e85d8741b1a0427)
Reviewed-on: https://chromium-review.googlesource.com/843061
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>