UPSTREAM: crypto: introduce crypto wait for async op
Invoking a possibly async. crypto op and waiting for completion
while correctly handling backlog processing is a common task
in the crypto API implementation and outside users of it.

This patch adds a generic implementation for doing so in
preparation for using it across the board instead of hand
rolled versions.

Signed-off-by: Gilad Ben-Yossef <gilad@benyossef.com>
CC: Eric Biggers <ebiggers3@gmail.com>
CC: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

BUG=b:151242596
TEST=run kvm-xfstests verity

(cherry picked from commit ada69a1639eca)

Change-Id: Ib5cb9127cfd45510795ce34dee1ba6006d442b9f
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2275786
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Eric Biggers <ebiggers@google.com>