UPSTREAM: drm/msm: Make the value of RB_CNTL (almost) generic
We use a global ringbuffer size and block size for all targets and
at least for 5XX preemption we need to know the value the RB_CNTL
in several locations so it makes sense to calculate it once and use
it everywhere.

The only monkey wrench is that we need to disable the RPTR shadow
for A430 targets but that only needs to be done once and doesn't
affect A5XX so we can or in the value at init time.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Rob Clark <robdclark@gmail.com>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I43069a6dfb6e6a838e7b39547894bc2961083ad6
Signed-off-by: Archit Taneja <architt@codeaurora.org>
(cherry picked from commit 4d87fc32dfd20ad0d4ffd5204f6e5d41d281e128)
Reviewed-on: https://chromium-review.googlesource.com/859215
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>